\name{MCI}
\alias{MCI}
\docType{data}

\title{ Sample of 75 Hospital Stays}

\description{
  Sample of 75 hospital for major cardiovascular interventions}

\usage{data(MCI)}

\format{
  A data frame with 75 observations on the following 6 variables.
  \describe{
    \item{\code{Sex}}{Gender: 1=Female, 2=Male}
    \item{\code{Age}}{Age in years}
    \item{\code{LOS}}{Length of stay (days)}
    \item{\code{TypAdm}}{Type of admission: 1=Emergency 0=on notification}
    \item{\code{Dest}}{Destination: 1=Home 0=another hospital}
    \item{\code{Cost}}{Cost (Swiss francs)}
  }
}

\examples{
data(MCI)
}

\keyword{datasets}
